Remove S3 bucket input
I think it's better to not let a -github role to give permissions to an
S3 bucket directly. The -github role is supposed to be a limited role
with only permissions to assume other roles that can do real things.
Create resources, update state etc.

If -github had a permissions to the TF state then it would have to be
created in the tf-states account. It's not clean because the tf-states
account should only create S3 buckets for the TF states and IAM roles to
manage them.

This is also a breaking change in a context of the 1.* versions.